requirements.txt
absl-py==0.9.0
astor==0.8.1
bcrypt==3.1.7
blinker==1.4
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Bcrypt==0.7.1
Flask-Ext==0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-User==0.6.21
Flask-WTF==0.14.3
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
jwt==0.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
nexmo==2.4.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
passlib==1.7.2
Pillow==7.0.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycryptodome==3.9.6
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.13
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.0
WTForms==2.2.1